Change ZnZ macro to accept optional argument
Changes definition of ZnZ macro to use optional argument instead of
always requiring to provide it. This is useful in cases such as \ZnZ{n},
where the n is reduntant.

With this commit, \ZnZ{n} can be written as \ZnZ, and \ZnZ{m} has to be
written using square brackets, so \ZnZ[m].